type fixes

+++ ElectronicObserver.KancolleApi.Types/ApiReqMission/ReturnInstruction/APIReqMissionReturnInstructionResponse.cs
@@ -5,5 +5,5 @@ public class ApiReqMissionReturnInstructionResponse
 	[JsonPropertyName("api_mission")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<int> ApiMission { get; set; } = new();
+	public List<long> ApiMission { get; set; } = new();
 }
